function called has duplicates
Write a function called has duplicates that takes a list and returns True if there is any
element that appears more than once. It should not modify the original list.
i) Write a function called remove duplicates that takes a list and returns a new list with
only the unique elements from the original. Hint: they don’t have to be in the same order.